src: remove installation instructions for images (#56)
The steps required to import this into openshift are not simple,
and with the development of nodeshift and rhoar in general, this
does not seem to be something we should worry about.

In this commit, I remove the OpenShift installation instructions
as well as the image-streams-centos7.json file. These things just
create additional complexity with questionable user benefit.