Merging of RHOAR and RHSCL into a single 'nodejs' imagestream
Just introducing a new 8-RHOAR tag. Think this is the best we
can do at the moment. It will be harder to introduce a new imagestream
name like rhoar-nodejs and then move to just 'nodejs'. I'd like to
start with what we want it to be. Also need to merge, since RHSCL also
has '6' and '4'. Perhaps we can just swap out the '8' tag later for
just RHOAR.