test: add test for exit status from signal handler
This commit adds a signal handler that handles SIGTERM which is the
signal that docker will send to process 1 when docker stop is called.
The signal handle will call process.exit(222) which is then check after
the container has been stopped.

Currently this test is skipped for when DEV_MODE=true as the container
will instead by killed and the return status will be 137.