remove scl usage and link git instead (#86)
* remove scl usage and link git instead

This commit removes the usage of scl enable and insteads links the git
version in the image. Using scl enable cause an issue where the npm
process was not running as process 1 and signals got ignored.

* test: add test for exit status from signal handler

This commit adds a signal handler that handles SIGTERM which is the
signal that docker will send to process 1 when docker stop is called.
The signal handle will call process.exit(222) which is then check after
the container has been stopped.

Currently this test is skipped for when DEV_MODE=true as the container
will instead by killed and the return status will be 137.