test: add NPM_CONFIG_TARBALL check
This commit adds a check to verify that the environment variable
NPM_CONFIG_TARBALL matches the target node version.
It also moves the check of the node version of the base image to the
start of the test to allow for early failure if the version don't match.

The motivation for this is that checking we don't check the
NPM_CONFIG_TARBALL value (which contains the version) the building of
the addon in the test can fail and finding the cause of this error is
not that easy.