build: add debuginfo image and documentation
This commit adds the building of a second image for debuginfo which is
intended to be used in an OpenShift environment. The image is used to
start a container in the same cluster as a running application and to
be able to remote debug the node application using gdb.

Documentation has been added to a new docs directory as a suggestion and
contains instructions to set this up.